Split cards.js into modules
