+++ games/cards/cards.html
@@ -48,7 +48,7 @@
     <script src="https://cdn.jsdelivr.net/npm/peerjs@1.5.2/dist/peerjs.min.js"></script>
     <script src="../../js/codeconnect.js"></script>
     <script src="../../js/player-settings.js"></script>
-    <script src="cards.js"></script>
+    <script type="module" src="js/main.js"></script>
     <script src="../../js/footer.js"></script>
 </body>
 </html>
